* replace int type by the correct custom type for pos, idx and size.
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@21336 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/mathed/MacroTable.cpp
@@ -78,7 +78,7 @@ void MacroData::expand(vector<MathData> const & args, MathData & to) const
 }
 
 
-int MacroData::optionals() const
+size_t MacroData::optionals() const
 {
 	return optionals_;
 }
